Text: avoid clipping of overflowing inline-block content
Inline blocks content can have negative margins and can
overflow their line height or width. It's usually not
an issue as it will just be drawn over previous or next
lines, except when the line is at top or bottom of page,
where it would be clipped to the page rect clip.
So, just drop the page rect clip when drawing them, so
they can overflow in the page margins.